similarity-engine/requirements.txt
numpy
fastapi
uvicorn
tensorflow
python-multipart
pillow
faiss-cpu